Updated dev dependecies, added infection, fixed test namespaces
Main changes in composer.json and phpunit.xml.dist
Can't install phpdocumentor/phpdocumentor, removed
