+++ composer.json
@@ -15,14 +15,14 @@
         "php": "^5.6.0|^7.0.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8.35|^5.4.0",
-        "phpdocumentor/phpdocumentor":"2.*",
-        "phpmd/phpmd": "2.*",
-        "sebastian/phpcpd": "2.*",
-        "phploc/phploc": "2.*",
-        "squizlabs/php_codesniffer": "^3.4.0",
-        "phpcompatibility/php-compatibility": "^9.0",
-        "dealerdirect/phpcodesniffer-composer-installer": "^0.5.0"
+        "infection/infection": "0.13.x-dev",
+        "phpunit/phpunit": "^8.4@dev",
+        "phpmd/phpmd": "dev-master",
+        "sebastian/phpcpd": "^4.1",
+        "phploc/phploc": "^5.0@dev",
+        "squizlabs/php_codesniffer": "^3.0@dev",
+        "phpcompatibility/php-compatibility": "dev-master",
+        "dealerdirect/phpcodesniffer-composer-installer": "dev-master"
     },
     "autoload": {
         "psr-4": {
@@ -45,6 +45,29 @@
             "classes/src/operations/multiply.php",
             "classes/src/operations/divideby.php",
             "classes/src/operations/divideinto.php"
+        ]
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Matrix\\Test\\": "unitTests/classes/src/"
+        },
+        "files": [
+            "unitTests/classes/src/functions/adjointTest.php",
+            "unitTests/classes/src/functions/antidiagonalTest.php",
+            "unitTests/classes/src/functions/cofactorsTest.php",
+            "unitTests/classes/src/functions/determinantTest.php",
+            "unitTests/classes/src/functions/diagonalTest.php",
+            "unitTests/classes/src/functions/identityTest.php",
+            "unitTests/classes/src/functions/inverseTest.php",
+            "unitTests/classes/src/functions/minorsTest.php",
+            "unitTests/classes/src/functions/traceTest.php",
+            "unitTests/classes/src/functions/transposeTest.php",
+            "unitTests/classes/src/operations/addTest.php",
+            "unitTests/classes/src/operations/directsumTest.php",
+            "unitTests/classes/src/operations/subtractTest.php",
+            "unitTests/classes/src/operations/multiplyTest.php",
+            "unitTests/classes/src/operations/dividebyTest.php",
+            "unitTests/classes/src/operations/divideintoTest.php"
         ]
     },
     "scripts": {
