+++ unitTests/classes/src/Operators/AdditionTest.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace Matrix\Operators;
+namespace Matrix\Test\Operators;
 
-use \Matrix\Matrix;
+use Matrix\Matrix;
 use Matrix\Exception;
-use \Matrix\BaseTestAbstract;
+use Matrix\Operators\Addition;
+use Matrix\Test\BaseTestAbstract;
 
 class AdditionTest extends BaseTestAbstract
 {
