+++ unitTests/classes/src/Operators/MultiplicationTest.php
@@ -2,9 +2,12 @@
 
 namespace Matrix\Operators;
 
-use \Matrix\Matrix;
+namespace Matrix\Test\Operators;
+
+use Matrix\Matrix;
 use Matrix\Exception;
-use \Matrix\BaseTestAbstract;
+use Matrix\Operators\Multiplication;
+use Matrix\Test\BaseTestAbstract;
 
 class MultiplicationTest extends BaseTestAbstract
 {
