+++ unitTests/classes/src/functions/antidiagonalTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Matrix;
+namespace Matrix\Test;
 
-use Matrix\Matrix as Matrix;
+use Matrix\Matrix;
 
 class antidiagonalTest extends BaseTestAbstract
 {
