+++ unitTests/classes/src/functions/identityTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Matrix;
+namespace Matrix\Test;
 
-use Matrix\Matrix as Matrix;
+use Matrix\Matrix;
 
 class identityTest extends BaseTestAbstract
 {
