+++ unitTests/classes/src/operations/addTest.php
@@ -1,19 +1,19 @@
 <?php
 
-namespace Matrix;
+namespace Matrix\Test;
 
-use Matrix\Matrix as Matrix;
+use Matrix\Matrix;
 
-class subtractionTest extends BaseTestAbstract
+class addTest extends BaseTestAbstract
 {
-    protected static $operationName = 'subtraction';
+    protected static $operationName = 'addition';
 
     /**
      * @dataProvider dataProvider
      */
-    public function testSubtractionFunction($expected, $value1, $value2)
+    public function testAdditionFunction($expected, $value1, $value2)
     {
-        $result = subtract($value1, $value2);
+        $result = add($value1, $value2);
 
         //    Must return an object of the correct type...
         $this->assertIsMatrixObject($result);
@@ -24,10 +24,10 @@ class subtractionTest extends BaseTestAbstract
     /**
      * @dataProvider dataProvider
      */
-    public function testSubtractionInvoker($expected, $value1, $value2)
+    public function testAdditionInvoker($expected, $value1, $value2)
     {
         $matrix = new Matrix($value1);
-        $result = $matrix->subtract($value2);
+        $result = $matrix->add($value2);
 
         //    Must return an object of the correct type...
         $this->assertIsMatrixObject($matrix);
@@ -41,7 +41,7 @@ class subtractionTest extends BaseTestAbstract
     {
         return [
             [
-                [[3, -2], [9, -4]],
+                [[-1, 6], [-3, 12]],
                 [[1, 2], [3, 4]], [[-2, 4], [-6, 8]],
             ],
         ];
