+++ unitTests/classes/src/operations/dividebyTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Matrix;
+namespace Matrix\Test;
 
-use Matrix\Matrix as Matrix;
+use Matrix\Matrix;
 
 class dividebyTest extends BaseTestAbstract
 {
